requirements.txt
PyYAML
Pillow
colormath
